Convert fractal demo to use user's project.
Follow the pattern of the other demos and stop using robot accounts with a hard coded project for the fractal demo. Along the way, fix up the need to have a custom image (it is optional and detected).  There is still a hard coded shared bucket for the assets needed in the VM.

Also remove unnecessary scope from the service account set up for the VM itself.  None of our demos require the VM to call the GCE API.