Distributed tile rendering.
The go server can now break big tiles down into smaller tiles and farm them out to other servers.  These are composited together and returned to the client.

To make this work, I had to lock down the scaling of the tiles so that they render the same thing no matter the tile size.

Moved to the pre-built go 1.1b3 runtime.  It loads faster and might have some features we'll need.